refactor: 승격/아웃박스 도메인을 domain.promotion 패키지로 분리
평평하던 domain 패키지에서 PromotionTask·OutboxStatus를 '함께 변하는 개념'끼리 묶어 domain.promotion으로 이동. 패키지 경계로 승격이라는 도메인 개념을 드러낸다.
